fixed recruiter dashboard css and removed side bar
