+++ dashboard/src/components/Footer.css
@@ -1,5 +1,5 @@
 .footer {
-    background-color: #2D445A;
+    background-color: #296e85;
     color: white;
     text-align: center;
     padding: 10px 0;
