+++ frontend/src/static/App.css
@@ -149,6 +149,7 @@ input:focus-visible {
   box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
   max-width: 1200px;
   margin: auto; /* Center the content */
+  margin-top: 10px;
 }
 
 .table {
